Safari-609.1.11.1/LayoutTests/accessibility/mac/aria-menu-item-selected-notification-expected.txt

Menu item 1
Menu item 2
Menu item 3
This tests that a AXMenuItemSelected notification gets fired when a menuitem is focused.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
Received menu item selected notification: AXMenuItemSelected
Menu item selected element: AXRole: AXMenuItem
Menu item selected element description: AXDescription: item1

Received menu item selected notification: AXMenuItemSelected
Menu item selected element: AXRole: AXMenuItem
Menu item selected element description: AXDescription: item2

Received menu item selected notification: AXMenuItemSelected
Menu item selected element: AXRole: AXMenuItem
Menu item selected element description: AXDescription: item3

PASS successfullyParsed is true

TEST COMPLETE
